requirements.txt
streamlit
streamlit-extras
langchain
tiktoken
openai
python-dotenv
pandas
pypdf2
faiss-cpu
